linebot-weather/requirements.txt
Flask
line-bot-sdk
requests
schedule